mptcp: Don't allow SOCK_ZEROCOPY
It's not tested and syzkaller found a leak. Let's just disable it for
now. If somebody wants to support this feature, please go ahead :)

Support for TCP ZEROCOPY was added in v4.14, so the faulting commit is
the merge.

Fixes: 42a0700e8199 ("Merge tag 'v4.14-rc7' into mptcp_trunk")
Signed-off-by: Christoph Paasch <cpaasch@apple.com>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
(cherry picked from commit 838e4c41050683a7c6482cad1d7c5ebb2313e126)
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>